Removed averageConvergencePoint and added plotting routine
Removed call to averageConvergencePoint in combined_sequences.py
Added separate generate_plot function.
